move assertion to utils
